requirements.txt
ipykernel==7.1.0
matplotlib==3.10.7
scikit-learn==1.7.2
setuptools==59.6.0
torchvision==0.24.1
